Refactor: #62 report파일 위치 변경
